+++ CHANGELOG.md
@@ -28,6 +28,7 @@
  * Bump fastjson2-extension-spring6 from 2.0.47 to 2.0.50
  * Bump hutool from 5.8.26 to 5.8.27
  * Bump oshi-core from 6.4.13 to 6.6.0
+ * Bump spring-boot-admin from 3.2.3 to 3.3.1
  * Add mybatis 3.5.16
  * Add commons-compress 1.26.1
  * Add mybatis-flex-spring-boot3-starter 1.8.8